\name{allGenes}
\alias{allGenes}
\docType{data}
\title{
The Whole Gene Set
}
\description{
The gene list where the drug signatures were identified from. 
}
\usage{data("allGenes")}
\format{
  The format is:
 chr [1:12992] "A1CF" "A2M" "A4GALT" "A4GNT" "AAAS" "AACS" "AADAC" "AAGAB" "AAK1" "AAMP" "AANAT" "AARS" "AARSD1" "AASDHPPT" ...
}

\examples{
data(allGenes)

}
\keyword{datasets}
