\name{iDOMO-package}
\alias{iDOMO-package}
\alias{iDOMO}
\docType{package}
\title{
Identification of Drug Combinations via Multi-Set Operations
}
\description{
The package is designed to prioritize the pair-wise drug combinations from a list of drugs. The input disease signature and drug induced signatures should be a list of gene symbols. 
}
\details{
Use the SET.Combn function to rank the drug combinations. 
}
\author{
Xianxiao Zhou <xianxiao.zhou@mssm.edu>
Maintainer: Xianxiao Zhou <xianxiao.zhou@mssm.edu>, Bin Zhang <bin.zhang@mssm.edu>
}
\references{
Xianxiao Zhou et al. iDOMO: Identification of Drug Combinations via Multi-Set Operations for Treating Complex Diseases. 
Minghui Wang, Yongzhong Zhao, and Bin Zhang (2015). Efficient Test and Visualization of MultiSet Intersections. Scientific Reports 5: 16923.
}

\keyword{ iDOMO }
\seealso{
\code{\link[<EMUDRA>:<EMUDRA>-package]{<EMUDRA>}}
}
