\name{TNBCsig}
\alias{TNBCsig}
\docType{data}
\title{
A Triple Negative Breast Cancer Signature
}
\description{
Up-regulated genes and down-regulated genes identified between the TNBC and matched adjecent normal samples.
}
\usage{data("TNBCsig")}
\format{
  The format is:
List of 2
 $ disUp: chr [1:1485] "GPR19" "FAM72B" "ULBP1" "SKA1" ...
 $ disDn: chr [1:1713] "TCEAL7" "ABLIM3" "BMP4" "DUSP1" ...
}

\examples{
data(TNBCsig)

}
\keyword{datasets}
