\name{Top20DrugSig}
\alias{Top20DrugSig}
\docType{data}
\title{
Signatures Induced by 20 Drugs
}
\description{
Up-regulated genes and down-regulated genes between drug- and vehicle-treated cell line samples.
}
\usage{data("Top20DrugSig")}
\format{
  The format is:
List of 20
 $ colforsin      :List of 2
  ..$ : chr [1:774] "DDR1" "CSRP1" "NDRG1" "GLUL" ...
  ..$ : chr [1:913] "RFC2" "FNTB" "TARDBP" "STARD7" ...
 $ 5109870        :List of 2
  ..$ : chr [1:864] "ESRRA" "TRADD" "FNTB" "GUK1" ...
  ..$ : chr [1:619] "DDX5" "GLUL" "DNAJB1" "SRSF2" ...
 $ etoposide      :List of 2
  ..$ : chr [1:207] "ABCA12" "ACAA1" "ACTA2" "ALAS1" ...
  ..$ : chr [1:276] "ADSL" "AGPAT5" "AIF1" "AK2" ...
 $ resveratrol    :List of 2
  ..$ : chr [1:305] "ABAT" "ABHD2" "ACAA1" "ACRV1" ...
  ..$ : chr [1:516] "ACTG2" "ADNP" "AIM1" "AKAP1" ...
 $ trifluridine   :List of 2
  ..$ : chr [1:72] "ADRBK2" "ALDH1A3" "ANKRA2" "AP3M2" ...
  ..$ : chr [1:84] "ARHGAP19" "ARL6IP1" "ARL6IP4" "ASPM" ...
 $ methotrexate   :List of 2
  ..$ : chr [1:295] "ABCC10" "ABCC5" "ABCD4" "ABHD4" ...
  ..$ : chr [1:428] "ABCE1" "ADD3" "ADNP" "AGFG1" ...
 $ quinostatin    :List of 2
  ..$ : chr [1:1023] "HSPA6" "KAT7" "TMEM59" "NDRG1" ...
  ..$ : chr [1:1399] "DDX5" "SMNDC1" "ATP6V0B" "DNAJB1" ...
 $ ciclopirox     :List of 2
  ..$ : chr [1:383] "ABCC1" "ABHD2" "ABHD4" "ADAM8" ...
  ..$ : chr [1:412] "ABHD10" "ADNP" "ADRB2" "AIM1" ...
 $ etacrynic acid :List of 2
  ..$ : chr [1:30] "AKR1C2" "CCNG2" "CLCN3" "FAS" ...
  ..$ : chr [1:39] "AKAP1" "AP4S1" "BDH1" "CTBP1" ...
 $ calmidazolium  :List of 2
  ..$ : chr [1:1364] "HSPA6" "CSRP1" "WARS" "DNAJB1" ...
  ..$ : chr [1:1711] "EPHB3" "GAS6" "TRADD" "BAD" ...
 $ pyrvinium      :List of 2
  ..$ : chr [1:949] "AAK1" "AARS" "ABCA12" "ABHD4" ...
  ..$ : chr [1:1030] "ABCG1" "ABHD10" "ABHD11" "ACAD8" ...
 $ monobenzone    :List of 2
  ..$ : chr [1:46] "ABHD2" "ATF3" "CCL13" "CCNE2" ...
  ..$ : chr [1:91] "ADD3" "ARHGAP19" "ARL6IP1" "ASPM" ...
 $ azacitidine    :List of 2
  ..$ : chr [1:728] "AARSD1" "ACAA2" "ACAT1" "ACLY" ...
  ..$ : chr [1:1516] "ABCC5" "ABCD3" "ABCE1" "ABCF1" ...
 $ cytochalasin B :List of 2
  ..$ : chr [1:139] "CSRP1" "CAP1" "AHCYL1" "FLNA" ...
  ..$ : chr [1:242] "STARD7" "ABCF1" "SMNDC1" "MLEC" ...
 $ puromycin      :List of 2
  ..$ : chr [1:1527] "AAK1" "AARS" "AATK" "ABAT" ...
  ..$ : chr [1:1646] "AACS" "AAMP" "ABCC10" "ABCD3" ...
 $ wortmannin     :List of 2
  ..$ : chr [1:234] "ABCG1" "ABHD2" "ACP6" "ADAM17" ...
  ..$ : chr [1:251] "ABCE1" "ABCF2" "ADO" "AK2" ...
 $ 0175029-0000   :List of 2
  ..$ : chr [1:2597] "A2M" "AAAS" "AACS" "AAMP" ...
  ..$ : chr [1:2493] "AASDHPPT" "AATF" "ABCA11P" "ABCF1" ...
 $ latamoxef      :List of 2
  ..$ : chr [1:737] "ABCC2" "ABCC9" "ABCD2" "ABP1" ...
  ..$ : chr [1:608] "ABCF2" "ABHD5" "ACACA" "ACAD8" ...
 $ rifabutin      :List of 2
  ..$ : chr [1:1110] "AAK1" "ABAT" "ABCC1" "ABCD3" ...
  ..$ : chr [1:959] "ABCC5" "ABCE1" "ABCF1" "ABL1" ...
 $ trifluoperazine:List of 2
  ..$ : chr [1:251] "ABHD2" "ACBD3" "ACLY" "ADAM17" ...
  ..$ : chr [1:228] "ABCE1" "ABCF2" "ADCK2" "AGK" ...
}

\examples{
data(Top20DrugSig)

}
\keyword{datasets}
